% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracerutils.R
\name{catmaidPlot}
\alias{catmaidPlot}
\title{Retrieve a neuron from CATMAID and plot it, along with one or more CATMAID volumes if desired}
\usage{
catmaidPlot(skid, volumes = NULL, ncol = NULL, vcol = NULL,
  valpha = NULL)
}
\arguments{
\item{skid}{Required; the skeleton ID of the neuron in CATMAID. Can accept a vector with multiple SKIDs.}

\item{volumes}{The names of the volumes in CATMAId to be plotted.  Defaults to NULL.}

\item{ncol}{The colour(s) to use when plotting the specified neuron(s).
If a vector of colour names is provided, the colours will be applied to the neurons in the order specified,
repeating from the beginning if there are more neurons than colours.
Uses default from NAT package if unspecified.}

\item{vcol}{The colour(s) to use when plotting the specified volume(s).
Behaves the same way as ncol.  Defaults to gray.}

\item{valpha}{The alpha value(s) to use when plotting volume(s).
Behaves the same way as ncol and vcol.  Defaults to 0.5.}
}
\description{
Retrieve a neuron from CATMAID and plot it, along with one or more CATMAID volumes if desired
}

