% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_duplicate_synapses}
\alias{check_duplicate_synapses}
\title{Check a neuron for potential duplicate synapses}
\usage{
check_duplicate_synapses(neuron = NULL, skid = NULL, xy_threshold = 200,
  z_threshold = 200, fileout = NULL)
}
\arguments{
\item{neuron}{A CATMAID \code{neuron} object; required if \code{skid} is not provided.}

\item{skid}{The skeleton ID of a neuron in CATMAID; required if \code{neuron} is not provided.  Will only be used if \code{neuron} is not provided.}

\item{xy_threshold}{The maximum X-Y distance (in the same units as your neuron) between two connectors for them to be considered potential duplicates.  Defaults to 200.}

\item{z_threshold}{The maximum Z distance (in the same units as your neuron) between two connectors for them to be considered potential duplicates.  Defaults to 200.}

\item{fileout}{Optional; the path to a CSV file where the result should be written.}
}
\value{
A data frame containing the connector ID of each connector in a pair, a CATMAID URL for each, the X-Y and Z distances between the connectors, and the skeleton ID of the neuron.
}
\description{
Given a neuron or skeleton ID, this function will look for outgoing connectors that are close enough to represent the same synapse.
The function will return a data frame (optionally written to a CSV) with IDs of the identified connector pairs, their distances, and URLs to the connectors in CATMAID.
}
\details{
Requires the \code{catmaid.server} option to be set in .Rprofile
}
