% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracerutils_data.R
\docType{data}
\name{glomerular.valence}
\alias{glomerular.valence}
\title{Functional valence of different antennal lobe glomeruli}
\format{A named character vector containing the functional valence of each antennal lobe glomerulus
 (as described in Paavo Huoviala's thesis).  Possible valences are "Attractive", "Aversive, "Pheromonal",
 "Egg_laying", "Food", and "Unknown".  Element names correspond to the names of glomeruli.}
\description{
Functional valence of different antennal lobe glomeruli
}

