% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracerutils.R
\name{quickNBLAST}
\alias{quickNBLAST}
\title{NBLAST a neuron in CATMAID (FAFB) against the full flycircuit neuron list}
\usage{
quickNBLAST(skid)
}
\arguments{
\item{skid}{The skeleton ID of the neuron in CATMAID}
}
\value{
The full NBLAST results object
}
\description{
NBLAST a neuron in CATMAID (FAFB) against the full flycircuit neuron list
}

