% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlesheets.R
\name{read_team_sheet}
\alias{read_team_sheet}
\alias{get_team_sheet}
\title{Read / access a google sheet from a Google team drive}
\usage{
read_team_sheet(path = NULL, id = NULL, team_drive = NULL, ws = 1,
  ...)

get_team_sheet(path = NULL, id = NULL, team_drive = NULL)
}
\arguments{
\item{path}{File name or other path specifier}

\item{id}{File id (quicker than using the name)}

\item{team_drive}{Anything that specifies the team drive}

\item{ws}{positive integer or character string specifying index or title,
respectively, of the worksheet. The special value of \code{Inf} means all
sheets.}

\item{...}{Additional arguments passed to
\code{\link[googlesheets]{gs_read_csv}}}
}
\value{
A \code{tibble} or, when \code{ws} specifies more that one sheet, a
  list of tibbles nammed by the sheet name. For \code{get_team_sheet} a
  \code{\link[googlesheets]{googlesheet}} object.
}
\description{
\code{read_team_sheet} reads (and finds) a google sheet on a
  team drive

\code{get_team_sheet} finds a google sheet on a team drive. You
  can use the returned
}
\details{
Google sheet objects on team drives are not simply accessible by the
  googlesheets package.
}
\examples{
\dontrun{
pns<-read_team_sheet('AL_fragments_spreadsheet', team_drive='flyconnectome',ws=1:2)
gs <- get_team_sheet(id='1QMCbGXmJqxIm2K-IvNivN3od9QKzZCb5kA_Dqn9xjNs')
}
}
