% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_sampling.R
\name{sample_connections}
\alias{sample_connections}
\title{Randomly sample synapses or connections of a neuron}
\usage{
sample_connections(neuron, number = NULL, type = c("downstream", "upstream",
  "connector"), fileout = NULL, node.id = NULL,
  node.direction = c("downstream", "upstream"), volume = NULL,
  plot = FALSE)
}
\description{
Randomly sample synapses or connections of a neuron
}

