% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tagged_nodes}
\alias{tagged_nodes}
\title{Generate URLs for specifically tagged nodes on a neuron}
\usage{
tagged_nodes(neuron = NULL, tag, node.id = NULL,
  node.direction = c("downstream", "upstream"), volume = NULL,
  fileout = NULL)
}
\arguments{
\item{neuron}{Required; a A CATMAID \code{neuron} object or skeleton ID.}

\item{tag}{Required; a string specifying a tag.}

\item{node.id}{Optional; the ID of a node at which to cut the neuron, if only a portion of the neuron should be considered.}

\item{node.direction}{If a cut node is provided, which portion of the cut neuron to retain.  Options are \code{downstream} (default) or \code{upstream}.}

\item{volume}{Optional; a string specifying a FAFB neuropil to filter the tagged nodes.}

\item{fileout}{Optional; the path to a CSV file where the result should be written.}
}
\value{
A subset of the neuron's treenode data frame (\code{d}) including only tagges nodes, with a CATMAID URL for each.
}
\description{
Given a neuron and a particular tag, this will return a data frame (optionally written to a CSV) with the node IDs, coordinates, and CATMAID URLs.
}
\details{
Requires the \code{catmaid.server} option to be set in .Rprofile
}
