% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_and_defunct.R
\name{tracerutils-defunct}
\alias{tracerutils-defunct}
\alias{quick_nblast}
\title{Removed functions}
\usage{
quick_nblast(skid)
}
\arguments{
\item{skid}{The skeleton ID of the neuron in CATMAID}
}
\value{
The full NBLAST results object
}
\description{
Removed functions

DEFUNCT - use \code{\link{nblast_fafb}}
}
\details{
Requires the \code{flycircuit.datadir} option to be set and the \code{nat.default.neuronlist} option to be set to "dps" in .Rprofile
}
