% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracerutils.R
\name{quick_nblast}
\alias{quick_nblast}
\title{NBLAST a CATMAID neuron against the flycircuit data set}
\usage{
quick_nblast(skid)
}
\arguments{
\item{skid}{The skeleton ID of the neuron in CATMAID}
}
\value{
The full NBLAST results object
}
\description{
NBLAST a neuron in CATMAID (FAFB) against the full flycircuit neuron list
}
\details{
Requires the \code{flycircuit.datadir} option to be set and the \code{nat.default.neuronlist} option to be set to "dps" in .Rprofile
}

