% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracerutils-package.R
\docType{package}
\name{tracerutils-package}
\alias{tracerutils-package}
\alias{_PACKAGE}
\alias{tracerutils}
\title{tracerutils: Utilities for EM Neuron Tracers}
\description{
Helpful functions for handling neurons traced from EM volumes.
    Interacts with CATMAID and NBLAST.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/flyconnectome/tracerutils}
  \item \url{https://flyconnectome.github.io/tracerutils}
  \item Report bugs at \url{https://github.com/flyconnectome/tracerutils/issues}
}

}
\author{
\strong{Maintainer}: Fiona Love \email{fl299@cam.ac.uk} (0000-0002-7805-5234)

Authors:
\itemize{
  \item Gregory Jefferis \email{jefferis@gmail.com} (0000-0002-0587-9355)
}

}
\keyword{internal}
