% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3D_plots.R
\name{plot_catmaid}
\alias{plot_catmaid}
\title{Quickly plot neurons from CATMAID}
\usage{
plot_catmaid(skid, volumes = NULL, ncol = NULL, vcol = NULL,
  valpha = NULL)
}
\arguments{
\item{skid}{Required; the skeleton ID(s) of the neuron(s) in CATMAID.
Any format accepted by \code{read.neurons.catmaid()} from the \code{catmaid} package can be used.}

\item{volumes}{The names of the volumes in CATMAID to be plotted.  Defaults to NULL.}

\item{ncol}{The colour(s) to use when plotting the specified neuron(s).
If a vector of colour names is provided, the colours will be applied to the neurons in the order specified,
repeating from the beginning if there are more neurons than colours.
Uses default from NAT package if unspecified.}

\item{vcol}{The colour(s) to use when plotting the specified volume(s).
Behaves the same way as ncol.  Defaults to gray.}

\item{valpha}{The alpha value(s) to use when plotting volume(s).
Behaves the same way as ncol and vcol.  Defaults to 0.1.}
}
\value{
Returns the neuron(s) retrieved from CATMAID
}
\description{
Retrieve one or more neurons from CATMAID and plot them, along with one or more CATMAID volumes if desired
}
\examples{
\dontrun{
plot_catmaid(skid = 'WTPN2017_uPN_right', volumes = 'v14.neuropil')
}
}
