% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_neuron_local}
\alias{split_neuron_local}
\title{Split a neuron and return 'downstream' section}
\usage{
split_neuron_local(skid = NULL, neuron = NULL, node,
  return = c("downstream", "upstream"))
}
\arguments{
\item{skid}{Required unless \code{neuron} is provided; the skeleton ID of the neuron to split (passed on to
\code{\link{catmaid_skids}})}

\item{neuron}{Required unless \code{skid} is provided; a \code{neuron} object to split}

\item{node}{Required; the ID of the node where the neuron should be split (or
a string naming a tag)}

\item{return}{Which part of the split neuron to return (\code{downstream} or \code{upstream}). Defaults to \code{downstream}.}
}
\value{
A \code{neuron} object representing the specified portion of the
  split neuron
}
\description{
Split a neuron at a particular node and return the 'downstream' section as a
new neuron object
}
\details{
This does not actually split the neuron in CATMAID; split is only performed
on the local neuron object. If a string is used to specify the split
\code{node} (rather than the CATMAID node id), it is matched against named
tags present in the input neuron. There will be an error if this does not
resolve to a single unique node.
}
\examples{
\donttest{
DL4.LH=split_neuron_local("glomerulus DL4 right", "SCHLEGEL_LH")
plot(DL4.LH)
}
}
