% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_neuron_local}
\alias{split_neuron_local}
\title{Split a neuron and return 'downstream' section}
\usage{
split_neuron_local(skid, node, return = "child")
}
\arguments{
\item{skid}{Required; the skeleton ID of the neuron to split}

\item{node}{Required; the ID of the node where the neuron should be split}

\item{return}{FIXME: currently ignored}
}
\value{
A \code{neuron} object representing the 'downstream' portion of the split neuron
}
\description{
Split a neuron at a particular node and return the 'downstream' section as a new neuron object
}
\details{
This does not actually split the neuron in CATMAID; split is only performed on the local neuron object.
}
