% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapses_per_neuropil.R
\name{synapses_per_neuropil}
\alias{synapses_per_neuropil}
\title{Count the number of synapses one or more neurons make in each neuropil volume}
\usage{
synapses_per_neuropil(skids = NULL, neurons = NULL, reference = c("FAFB",
  "FCWB"))
}
\arguments{
\item{skids}{A list of skeleton IDs; required if \code{neurons} is not provided.  Will only be used if \code{neurons} is not provided.}

\item{neurons}{A neuron list; required if \code{skids} is not provided.}
}
\value{
Returns a list containing one data frame for each neuron/skeleton ID provided.
Each data frame will contain a row for each neuropil, with two columns indicating the number of outgoing and incoming synapses that the neuron has in that region.
}
\description{
Count the number of synapses one or more neurons make in each neuropil volume
}
